Updates the module to automatically preconfigure the node
