+++ scripts/preconfig/Cargo.toml
@@ -6,12 +6,8 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [[bin]]
-name = "funding"
-path = "src/funding.rs"
-
-[[bin]]
-name = "nft_sale"
-path = "src/nftsale.rs"
+name = "preconfig"
+path = "src/preconfig.rs"
 
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
