+++ scripts/preconfig/src/lib.rs
@@ -1,4 +1,1 @@
-mod nftsale;
-
-#[cfg(test)]
-mod funding;
+mod preconfig;
